requirements.txt
langchain==0.3.0
langgraph==0.2.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.50.0
azure-search-documents==11.4.0
azure-identity==1.17.0
semantic-kernel==1.11.0
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
streamlit==1.39.0
plotly==5.24.0
pandas==2.2.0
python-dotenv==1.0.0
requests==2.32.0
httpx==0.27.0
